# Conflicts:
#	sys/flexyadmin/assets/js/components/form/flexy-form.vue


Former-commit-id: c9d3b1da1e8adf855884f848dae6851716085efd
Former-commit-id: 3a6727c3fe1f91eb55c706b5da70eb6414b32185 [formerly d103076700338c57667abf910c9c8d7592a84f24]
Former-commit-id: 75209d76ff9625a9924a29aa6b8560d5a01fb1fc

+++ sys/flexyadmin/models/data/Data_Core.php
@@ -2769,10 +2769,12 @@ Class Data_Core extends CI_Model {
     $cache_filter = 'data_result_';
     $cached_results = $this->cache->cache_info();
     foreach ($cached_results as $cache) {
-      foreach ($this->settings['cache_group'] as $filter) {
-        $filter = $cache_filter.$filter;
-        if ( substr($cache['name'],0,strlen($filter))===$filter ) {
-          $this->cache->delete($cache['name']);
+      if ($this->settings['cache_group']) {
+        foreach ($this->settings['cache_group'] as $filter) {
+          $filter = $cache_filter.$filter;
+          if ( substr($cache['name'],0,strlen($filter))===$filter ) {
+            $this->cache->delete($cache['name']);
+          }
         }
       }
     }
